Cleared errors related to previous commit
Code for the get_range function was removed previously. Code is re-added into the function
Derivative function was adding into the integral variable, incorrect code was fixed.
Currently, the code is continuously turning the car which causes it to crash into the wall preventing it to move any further.
All parts of the code appears to be otherwise functional.

TODO: pid_control should return it's value for scan_callback to evaluate. If this value is outside of acceptable ranges, the call to change it's steering angle will be done. Otherwise, continue to drive at specified speed.