Changed best gap to deepest gap
Best gap is now determined by the one that is the deepest
Fixed crash which occurs at weird points. Crash was caused by a for loop that was going out of bounds.
This for loop has been commented out.
Loop may not be needed as the algorithm targets the center of the deepest gap anyways.