Added waypoints from levine hall
I kept the first point at 0 but maybe we should get rid of this since this is the starting location and we aren't driving to it? another thing is that once a full lap is complete should we get rid of the values after? I think we should get rid of the values at line 250 since they went from negative to positive indicating a full lap is complete. Let me know!