File runs
Moved all of the currently used parameters to be declared and retrieved inside the constructor.
To use these values, (self.___) will have to be used. For the sake of readability.
File has been cleared of python errors. Currently, the car just drives and fails to turn anywhere.
It does not seem to be doing much with the way point it was provided.