Iterative way points
Included a new function that will search for the next iterative waypoint instead of just the furthest one.

This Implementation only works clockwise. Some more consideration needed so that it can work in both angles.

NOTE: This implementation attempts to find the closest way point at every scan. If there are overlapping way points, it may pick the wrong one.
Possible solution is on line 341, pick the next iterative one instead of just the middle.