Adjustable speed
Original issue was the steering angle is in radians but the calculations are done in degrees. This resulted in only the top level if statement being the only one being executed.
Added some functions that should change the look ahead but they don't seem to be very helpful. Commented out for now